pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "paper-summarizer"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "docling>=2.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pypdf>=4.0.0",
    "tqdm>=4.66.0",
]

[project.scripts]
summarize-papers = "summarizer.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["summarizer*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
python_classes = ["Test*"]
addopts = "-v --tb=short"
pythonpath = ["."]
markers = [
    "integration: marks tests that require real external resources (PDFs, LM Studio)",
]
